requirements.txt
requests>=2.31.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
xgboost>=2.0.0
SQLAlchemy>=2.0.0
